requirements.txt
pillow<8.2
